fix(controller): modify webhook generation
In pre version, secret was generated by controller, so that, after uninstallation, the secret will
still reamin in k8s. When install next time, contoller cannot not create a new secret, it will occur
an error.

This commit will get and check if someone has updated secret, or the
controller will update it.